requirements.txt
psycopg2-binary~=2.9.9
SQLAlchemy~=2.0.34
python-dotenv~=1.0.1
aiogram~=3.12.0
uvicorn~=0.30.6
starlette~=0.38.5
bcrypt~=4.2.0
itsdangerous~=2.2.0
starlette-admin~=0.14.1
